Add simple shell and partitioned image creation
