+++ compile_tools.py
@@ -53,14 +53,15 @@ def compile_kernel():
     os.makedirs('build', exist_ok=True)
     run([NASM, '-f', 'elf32', 'src/boot.s', '-o', 'build/boot.o'])
     run([CC, '-m32', '-ffreestanding', '-O2', '-c', 'src/kernel.c', '-o', 'build/kernel.o'])
+    run([CC, '-m32', '-ffreestanding', '-O2', '-c', 'src/shell.c', '-o', 'build/shell.o'])
     run([CC, '-m32', '-ffreestanding', '-O2', '-c', 'src/string.c', '-o', 'build/string.o'])
     run([CC, '-m32', '-ffreestanding', '-O2', '-c', 'src/fs/vfs.c', '-o', 'build/vfs.o'])
     run([CC, '-m32', '-ffreestanding', '-O2', '-c', 'src/fs/ext2.c', '-o', 'build/ext2.o'])
     run([CC, '-m32', '-ffreestanding', '-O2', '-c', 'src/fs/fat32.c', '-o', 'build/fat32.o'])
     run([CC, '-m32', '-ffreestanding', '-O2', '-c', 'src/fs/ntfs.c', '-o', 'build/ntfs.o'])
     run([LD, '-m', 'elf_i386', '-T', 'linker.ld', '-o', 'build/kernel.bin',
-         'build/boot.o', 'build/kernel.o', 'build/string.o', 'build/vfs.o',
-         'build/ext2.o', 'build/fat32.o', 'build/ntfs.o'])
+         'build/boot.o', 'build/kernel.o', 'build/shell.o', 'build/string.o',
+         'build/vfs.o', 'build/ext2.o', 'build/fat32.o', 'build/ntfs.o'])
 
 def compile_bootloader(kernel_sectors, rootfs_sectors, rootfs_size):
     args = [NASM, '-f', 'bin',
@@ -82,17 +83,31 @@ def make_image():
     compile_bootloader(ksectors, rsectors, rsize)
 
     disk_size = 16 * 1024 * 1024  # 16MB virtual hard disk
+    with open('build/bootloader.bin', 'rb') as inp:
+        mbr = bytearray(inp.read())
+
+    # build partition table (MBR)
+    def part_entry(boot, ptype, start, count):
+        e = bytearray(16)
+        e[0] = 0x80 if boot else 0x00
+        e[4] = ptype
+        e[8:12] = struct.pack('<I', start)
+        e[12:16] = struct.pack('<I', count)
+        return e
+
+    mbr[446:462] = part_entry(True, 0x0B, 1, ksectors)
+    mbr[462:478] = part_entry(False, 0x83, 1 + ksectors, rsectors)
+
     with open('OptrixOS.img', 'wb') as out:
         out.truncate(disk_size)
         out.seek(0)
-        with open('build/bootloader.bin', 'rb') as inp:
-            out.write(inp.read())
-        out.seek(512)
-        with open('build/kernel.bin', 'rb') as inp:
-            out.write(inp.read())
-        out.seek(512 + ksectors * 512)
-        with open('rootfs.bin', 'rb') as inp:
-            out.write(inp.read())
+        out.write(mbr)
+        out.seek(1 * 512)
+        with open('build/kernel.bin', 'rb') as k:
+            out.write(k.read())
+        out.seek((1 + ksectors) * 512)
+        with open('rootfs.bin', 'rb') as rf:
+            out.write(rf.read())
 
 def make_iso():
     run([MKISOFS_EXE, '-quiet', '-o', 'OptrixOS.iso', '-b', 'OptrixOS.img',
