+++ src/shell.h
@@ -1,0 +1,6 @@
+#ifndef SHELL_H
+#define SHELL_H
+
+void shell_run(void);
+
+#endif /* SHELL_H */
